windows_requirements.txt
torch
jupyterlab
matplotlib
seaborn
tqdm
pylint
autopep8
pytest
python-dotenv
python-dotenv[cli]
msedge-selenium-tools
selenium==3.141
windows-curses